% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lslope_test_paired.R
\name{coef.lslope.test}
\alias{coef.lslope.test}
\title{Coef Method for lslope.test Objects}
\usage{
\method{coef}{lslope.test}(object, ...)
}
\arguments{
\item{object}{An object of class "lslope.test".}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A named numeric vector of lslope point estimates.
}
\description{
Extracts lslope point estimates from a lslope.test object.
}
