% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refit_rdgraph_regression.R
\name{compute.filter.weights.matrix}
\alias{compute.filter.weights.matrix}
\title{Compute filter weights for all eigenvalues and all eta candidates}
\usage{
compute.filter.weights.matrix(eigenvalues, eta.grid, filter.type)
}
\arguments{
\item{eigenvalues}{Vector of eigenvalues}

\item{eta.grid}{Vector of eta values}

\item{filter.type}{Filter type string}
}
\value{
Matrix of filter weights (m x length(eta.grid))
}
\description{
Compute filter weights for all eigenvalues and all eta candidates
}
\keyword{internal}
