% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riem_dcx_regression.R
\name{fitted.knn.riem.fit}
\alias{fitted.knn.riem.fit}
\title{Extract Fitted Values from kNN Riemannian Regression Fit}
\usage{
\method{fitted}{knn.riem.fit}(object, ...)
}
\arguments{
\item{object}{Object of class \code{knn.riem.fit}}

\item{...}{Additional arguments (currently unused)}
}
\value{
Numeric vector of fitted values (length n)
}
\description{
Extract Fitted Values from kNN Riemannian Regression Fit
}
