% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riem_dcx_regression.R
\name{get.rcx.optimal.k}
\alias{get.rcx.optimal.k}
\title{Extract Optimal k from Diagnostics Data Frame}
\usage{
get.rcx.optimal.k(diag.df)
}
\arguments{
\item{diag.df}{Data frame returned by k.diagnostics.plots}
}
\value{
A list with optimal k value and its index
}
\description{
Extract Optimal k from Diagnostics Data Frame
}
