% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulogit.R
\name{print.eigen.ulogit}
\alias{print.eigen.ulogit}
\title{Print Method for eigen.ulogit Objects}
\usage{
\method{print}{eigen.ulogit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"eigen.ulogit"}}

\item{...}{Additional arguments (currently ignored)}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for eigen.ulogit Objects
}
\examples{
fit <- list(
  call = quote(eigen.ulogit(x, y, fit.quadratic = TRUE)),
  model = "quadratic",
  beta = c(0.5, 1.2, -0.3),
  converged = TRUE,
  iterations = 5L,
  loglik = -12.34,
  aic = 30.68,
  bic = 34.12,
  errors = c(0.09, 0.11, 0.10)
)
class(fit) <- "eigen.ulogit"

fit
}
